Add files via upload
A Client has approached with an objective of establishing a reporting tool that provides insights into the performance of their two Power Plants. After thoroughly understanding the Clients requirements, I have established a connection with their database and conducted necessary data cleaning. I then utilised Microsoft Power BI to develop a comprehensive solution in the form of an interactive dashboard. The key features of the dashboard include:
1.	Energy Export to Grid
•	“Card” visuals are utilized to display Net Delivered Energy and Wind Speed on a user defined interval such as Daily, Monthly, or Yearly.
•	A "Bar" visual highlights the electricity generated by both power plants on a monthly basis.
•	Two "Treemaps" provide a quick understanding of the month that significantly contribute to the company's revenue.
•	A "Donut" visual displays the proportion of energy generated by both power plants.

2.	Capacity Factor tracking
Client wanted to track the Capacity Factor of both power plants, so I used "Gauge" and "Card" visuals to achieve this. It provides quick insights into:
•	Capacity Factor (Export Energy only)
•	Capacity Factor (sum of Export Energy and Non-Project Missed Volume) 
I also created a "Measure" to calculate "Energy required to achieve Capacity Factor," which is displayed using a Card visual.

3.	Wind Turbine Alarm History
To understand common alarms and take appropriate measures, I created a dashboard to analyze the history of alarms occurring on wind turbines.
•	A “Bar chart” visually highlights the number of occurrences of different alarms in relation to wind turbines.
•	A “Scatter plot” enables the user to analyze recurring alarms day-wise.
•	A “Donut chart” displays the alarm density on a monthly basis.
•	“Slicers” were created to allow the user to filter out alarms based on Year, Month, Date, and Alarm Code. 

4.	Preventive Maintenance Tracking
The client wanted to get an insight into the Maintenance Schedule of Power Plant and to become aware whenever allowable maintenance hours are breached, so I created a dashboard to display this information: 
•	Maintenance Hours of each turbine can be viewed on bar-chart based on time intervals (daily or monthly), maintenance type (3M or 6M), and turbine number.
•	A “Donut chart” displays the proportion of maintenance hours in relation to maintenance type.
•	A “Treemap” provides a quick understanding of the months that have significant amounts of downtime for maintenance purposes.
•	A "Gauge" visual displays the total downtime hours of the power plant

5.	Financial Analysis 
Financial analysis is a crucial tool for an organisation to make informed decisions, evaluate performance, and ensure financial stability and growth. I created a dashboard to display the company's financial performance.